pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "litra-control"
version = "0.1.0"
description = "Individual device control for Logitech Litra lights"
readme = "README.md"
requires-python = ">=3.6"
dependencies = ["pyusb"]

[tool.setuptools.packages.find]
where = ["."]
include = ["litra_control*"]

